+++ Runtime/ETA/Item.cs
@@ -1,4 +1,3 @@
-#nullable enable
 using System.IO;
 using ETA_Dependencies.Unity;
 using UnityEngine;
@@ -20,28 +19,32 @@ namespace ETA
         public string adUnitId = null!; //must be set in Unity Editor
         public bool allowImpression = true;
         public bool loadOnStart = true;
+        public bool interactable = false;
+        public bool enableRefresh = true;
         public float refreshTime = 10.0f;
         
-        private bool _isImpressed;
-        private float _afterImpressedTime;
+        private bool _startRefresh;
+        private float _refreshWaited;
         
 
         internal void Awake() // todo change Destroy process
         {
-            if(EtaSdk.OnceInitialized == false)
+            if(EasterAdSdk.OnceInitialized == false)
             {
-                EtaSdk.ItemAwakeQueue.Enqueue(this);
+                EasterAdSdk.ItemAwakeQueue.Enqueue(this);
                 EnableSDK();
                 return;
             }
             
-            if (EtaSdk.Instance.GetItemClient(adUnitId) != null)
+            if (EasterAdSdk.Instance.GetItemClient(adUnitId) != null)
             {
                 InstanceManager.DebugLogger.Log("Item already exist: " + adUnitId);
             }
 
             _client = GetClient(gameObject, adUnitId);
-            EtaSdk.Instance.AddItemClient(adUnitId, ref _client);
+            _client.AllowImpression = allowImpression;
+            _client.Interactable = interactable;
+            EasterAdSdk.Instance.AddItemClient(adUnitId, ref _client);
             InstanceManager.DebugLogger.Log("Item added: " + adUnitId);
         }
 
@@ -52,32 +55,43 @@ namespace ETA
         
         private void Update()
         {
-            if(_isImpressed)
+            _client.AllowImpression=allowImpression;
+            _client.Interactable=interactable;
+            
+            if (!enableRefresh) { return; }
+            
+            if(_startRefresh)
             {
-                _afterImpressedTime += Time.unscaledDeltaTime;
-                if (_afterImpressedTime >= refreshTime && Client.GetStatus() == ItemStatus.Impressed)
+                _refreshWaited += Time.unscaledDeltaTime;
+                if (_refreshWaited >= refreshTime)
                 {
-                    _isImpressed = false;
-                    _afterImpressedTime = 0.0f;
+                    _startRefresh = false;
+                    _refreshWaited = 0.0f;
                     Load();
                 }
+
+                if (Client.GetStatus() != ItemStatus.Impressed && Client.GetStatus() != ItemStatus.Interacted)
+                {
+                    _startRefresh = false;
+                    _refreshWaited = 0.0f;
+                }
             }
-            else if (Client.GetStatus() == ItemStatus.Impressed)
+            else if (Client.GetStatus() == ItemStatus.Impressed || Client.GetStatus() == ItemStatus.Interacted)
             {
-                _isImpressed = true;
-                _afterImpressedTime = 0.0f;
+                _startRefresh = true;
+                _refreshWaited = 0.0f;
             }
             else if (Client.GetStatus() == ItemStatus.Impressing)
             {
-                _isImpressed = false;
-                _afterImpressedTime = 0.0f;
+                _startRefresh = false;
+                _refreshWaited = 0.0f;
             }
         }
 
         private void OnDestroy()
         {
             try{
-                EtaSdk.Instance.RemoveItemClient(adUnitId);
+                EasterAdSdk.Instance.RemoveItemClient(adUnitId);
             }
             catch
             {
@@ -130,6 +144,9 @@ namespace ETA
         // {
         //     Client.Destroy();
         // }
+        
+        public abstract string StartInteraction();
+        public abstract void EndInteraction();
 
         /// <summary>
         /// <para xml:lang="ko">상속된 클래스에서 구현해야 합니다.</para>
@@ -149,7 +166,7 @@ namespace ETA
             
             if (easterAdEnabled)
             {
-                EtaSdk.CreateEtaSdk();
+                EasterAdSdk.CreateEtaSdk();
             }
         }
     }